backend/pyproject.toml
[tool.poetry]
name = "minesweeper"
version = "0.1.0"
description = ""
authors = ["Don Nguyen <ndon55555@gmail.com>", "Cleverina Cong <cccongx@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
pygame-minesweeper-core = "^1.0.18"
Flask-SocketIO = "^5.0.1"
gunicorn = "^20.1.0"
logdecorator = "^2.2"
SQLAlchemy = "^1.4.40"
PyMySQL = "^1.0.2"
eventlet = "0.30.2"
MarkupSafe = "2.0.1"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
flake8 = "^3.9.0"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
